requirements.txt
plotly==4.1.0
dash_html_components==1.0.3
decorator==4.4.0
dill==0.3.2
pandas==1.2.0
tqdm==4.51.0
lxml==4.6.2
requests==2.25.1
pymssql==2.1.5
dash_table==4.6.2
dash_core_components==1.9.1
dash==1.11.0
jsonpickle==2.0.0
pytest==6.2.4
PyYAML==5.4.1
responses==0.13.3
sphinx_rtd_theme==0.5.2
